Synced held map item data and rendering (#6)
Syncs the held map item data to spectators, and fixes arm rendering when
the spectated player is holding a map (previously the arms were
invisible).

On Fabric servers, this implementation sends an initial map data packet
when the player starts spectating. Then, as any map update packets are
sent to the target, the same packets are also sent to spectators.

On Paper servers, this is not possible to do without lots of reflection.
Listening to packets sent to targets would also be needed unless the
plugin also reimplemented how the map diff is sent. So instead, full map
renders are sent to all spectators every 30 seconds.