Add keybinds to teleport to players (#7)
Adds 3 customizable keybinds:
* Closest Player: Teleports to the closest player. If
no players are within tracking distance, the action will fail.
* Previous Player and Next Player:
Teleports to a player from the connected players list. Keeps track of
the last player teleported to and will adjust who is next selected based
on that. If no players are online, the action will fail.

If the existing "Spectate when teleporting" config option is enabled,
the user will automatically start spectating the player when using the
keybinds.

Adds a new "Keybinds open spectator menu" config option that when
enabled will automatically open the spectator hotbar menu and select the
player which has been teleported when using the keybinds.

The keybinds have no effect when used outside of spectator mode.